requirements.txt
asgiref==3.11.1
Django==5.2.13
gitdb==4.0.12
GitPython==3.1.46
python-dotenv==1.2.2
smmap==5.0.3
sqlparse==0.5.5
